AddTrade2

+++ mvcGCards/Models/TradeSenderView.cs
@@ -2,7 +2,7 @@
 {
     public class TradeSenderView
     {
-        public required int TradeId { get; set; }
+        public required long TradeId { get; set; }
         public required List<TradeCard> TradeCards { get; set; }
     }
 }
